// src/modules/venelux/domain/types/create-venelux-detail.type.ts
export interface CreateVeneluxDetail {
  solicitudnumero: string;
  itemnumero: string;
  codigomaterial: string;
  descripcionmaterial: string;
  coduni: string;
  unidadmedida: string;
  linea: string;
  sublinea: string;
  categoria: string;
  cantidadsolicitada: number;
  precioventa: number;
  observacion: string;
  materialnuevo: string;
}