// src/modules/venelux/domain/types/create-venelux-header.type.ts
export interface CreateVeneluxHeader {
  solicitudnumero: string;
  empresa: string;
  codigoobra: string;
  descripcionobra: string;
  numerocontrol: string;
  solicitanteuser: string;
  solicitantecodigo: string;
  fechasolicitud: string;
  fechautilizacion: string;
  observacion: string;
  actividad: string;
  direccionentrega: string;
  registradopor: string;
  owneruser: string;
}