// src/modules/venelux/domain/types/venelux-material.type.ts
export interface VeneluxMaterial {
  codigo: string;
  material: string;
  coduni: string | null;
  nroparte: string | null;
  codbarra: string | null;
  unidad: string | null;
  linea: string | null;
  sublinea: string | null;
  categoria: string | null;
  precio: number | null;
  codart: number | null;
  marca: string | null;
  noparte: string | null;
  imagen1: string | null;
  imagen2: string | null;
  imagen3: string | null;
  
}